Framework-client/requirements/common.txt
requests==2.12.5
psutil==5.3.1
pypiwin32==219
lxml==3.7.3
selenium==3.9.0
robotframework==3.0.2
